web/requirements.txt
aesedb==0.1.6
aiosmb==0.4.10
aiowinreg==0.0.10
alembic==1.12.0
amqp==5.1.1
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.3.0
arjun==2.2.1
asgiref==3.7.2
asn1crypto==1.5.1
asyauth==0.0.20
asysocks==0.2.12
azure-common==1.1.28
beautifulsoup4==4.12.3
bidict==0.23.1
billiard==4.1.0
blinker==1.6.2
bs4==0.0.1
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
colorclass==2.2.2
colored==2.2.3
crossplane==0.5.8
cryptography==42.0.5
dank==0.0.23
DateTime==5.3
datrie==0.8.2
Deprecated==1.2.14
docopt==0.6.2
dsinternals==1.2.4
editdistance==0.6.2
EditorConfig==0.12.3
fake-useragent==1.3.0
filelock==3.13.1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
flask-paginate==2023.10.24
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.3.2
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
future==1.0.0
google==3.0.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.4
idna==3.6
impacket==0.11.0
importlib-resources==6.1.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsbeautifier==1.14.11
Mako==1.2.4
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
minidump==0.0.23
minikerberos==0.4.4
mmh3==3.0.0
msldap==0.5.10
netaddr==0.7.20
openai==1.14.2
ordered-set==4.1.0
oscrypto==1.3.0
packaging==23.2
passlib==1.7.4
platformdirs==3.11.0
pluggy==1.3.0
pluginbase==1.0.1
prompt-toolkit==3.0.39
psutil==5.9.6
pyasn1==0.6.0
pybind11==2.11.1
pycparser==2.21
pycryptodomex==3.20.0
pycurl==7.45.3
pydantic==2.6.4
pydantic_core==2.16.3
pyfiglet==1.0.2
Pygments==2.16.1
PyMySQL==1.1.0
pyOpenSSL==24.0.0
pyparsing==3.1.2
pypykatz==0.6.9
PySocks==1.6.8
python-dateutil==2.8.2
python-engineio==4.9.0
python-slugify==8.0.1
python-socketio==5.11.1
PyYAML==6.0.1
requests==2.31.0
requests-file==1.5.1
requests-futures==1.0.1
rich==13.6.0
sentry-sdk==1.33.1
shodan==1.28.0
simple-websocket==1.0.0
six==1.16.0
snallygaster==0.0.12
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.21
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
tabulate==0.9.0
termcolor==2.3.0
text-unidecode==1.3
tldextract==5.1.1
tqdm==4.66.2
typing_extensions==4.8.0
tzdata==2023.3
unicrypto==0.0.10
urllib3==2.2.1
urlparse3==1.1
vine==5.0.0
virtualenv==20.24.6
wcwidth==0.2.10
webencodings==0.5.1
Werkzeug==3.0.1
WTForms==3.0.1